pyproject.toml
[project]
name = "anchorna"
dynamic = ["version"]
authors = [
    {name = "Tom Eulenfeld"},
]
description = "Find anchors in RNA/DNA sequences"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = [
    "bioinformatics", "RNA", "DNA", "sequence", "anchors"
]
dependencies = [
  "rnajena-sugar",
  "tqdm",
]

[project.urls]
"Homepage" = "https://github.com/rnajena/anchorna"
"Bug Tracker" = "https://github.com/rnajena/anchorna/issues"

[project.scripts]
anchorna = "anchorna.cli:run_cmdline"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "anchorna.__version__"}

[tool.setuptools.packages.find]
include = ["anchorna*"]

[tool.setuptools.package-data]
"anchorna.tests.data" = ["*.*"]
